Refactor of lab1: - add Graph class encapsulating dependencies - update CMakeLists.txt - clean up main.cpp
